Add files via upload
Previously, the app displayed inaccurate daily minimum and maximum temperatures. These fields are replaced with the forecasted temperatures and conditions at each of the next three times that are multiples of three hours per GMT. For example, if the time zone is Eastern Standard Time and the current time is 12:00, forecasts for 13:00, 16:00, and 19:00 will be displayed.

This update also adds detail to the weather conditions and makes some adjustments to the user interface.